se borrraba la foto por default. ahora se supone que no

+++ unAventonApp/views.py
@@ -186,7 +186,8 @@ def upload_foto(request):
         usuario = request.user.usuario
         file = request.FILES['files']
         if str(file).lower().endswith(('.jpg', '.png', '.jpeg', '.gif',)) == True:
-            usuario.foto_de_perfil.delete()
+            if not str(usuario.foto_de_perfil.name).count('default-user.png'):
+                usuario.foto_de_perfil.delete()
             usuario.foto_de_perfil = file
             usuario.save()
         return redirect('miPerfil')
